+++ backend/interfaces/api/server/router/setRouters.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/yoshihiro-shu/draft-backend/auth"
 	"github.com/yoshihiro-shu/draft-backend/controllers/api"
-	"github.com/yoshihiro-shu/draft-backend/handler"
+	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server/handler"
 	"github.com/yoshihiro-shu/draft-backend/request"
 )
